Rename errors.py to middleware_errors.py
The module holds the errors the middleware layer raises during call setup, not
errors in general, and the bare name invited unrelated things being put there.
